+++ tests/datasets/test_experiments.py
@@ -8,7 +8,7 @@ from phoenix.datasets.evaluators import (
     ContainsKeyword,
     HelpfulnessEvaluator,
 )
-from phoenix.datasets.evaluators.utils import create_evaluator
+from phoenix.datasets.evaluators.base import create_evaluator
 from phoenix.datasets.experiments import run_experiment
 from phoenix.datasets.types import (
     AnnotatorKind,
